# refactor: simplify RunScorePlugins for readability + performance