@pacoxu 2023-07-19
https://perf-dash.k8s.io/#/?jobname=kubemark-5000Nodes-scheduler&metriccategoryname=Scheduler&metricname=SchedulingLatency&Operation=Score&TestName=pod-anti-affinity

![image](https://github.com/kubernetes/kubernetes/assets/2010320/a229e9dd-d9ac-4b94-bdb4-a8eee5b9a9f8)

I saw a little performance downward in the performance dashboard. This seems to be expected.
/cc @alculquicondor @kerthcet